+++ system/modules/isotope/CHANGELOG.md
@@ -23,6 +23,7 @@ Version 1.3.0 stable (2012-??-??)
 - Default county is not set in Address Book edit mode for new address entry form
 - Fixed warning when product was not assigned to any category (#2998)
 - The back button in shop config list did not work
+- Address book module did not correctly handle default address checkbox (#3000)
 
 
 Version 1.3.rc2 (2012-03-19)